Added utility method to get custom cargo icon. Moved WheelRotationViaAnimation to another namespace.
